Implement metadata object
