+++ docs/examples.rst
@@ -67,3 +67,4 @@ Here is a list of example notebooks to illustrate how to use earthkit-data.
     :glob:
 
     examples/projection.ipynb
+    examples/numpy_fieldlist.ipynb
