+++ earthkit/data/core/fieldlist.py
@@ -1,0 +1,831 @@
+# (C) Copyright 2020 ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+#
+
+from abc import abstractmethod
+from collections import defaultdict
+
+from earthkit.data.core import Base
+from earthkit.data.core.index import Index
+from earthkit.data.decorators import cached_method
+from earthkit.data.utils.metadata import metadata_argument
+
+
+class Field(Base):
+    r"""Represents a Field."""
+
+    @property
+    @abstractmethod
+    def values(self):
+        r"""ndarray: Gets the values stored in the field as a 1D ndarray."""
+        self._not_implemented()
+
+    @property
+    @abstractmethod
+    def _metadata(self):
+        r"""Metadata: Gets the object representing the field's metadata."""
+        self._not_implemented()
+
+    def to_numpy(self, flatten=False, dtype=None):
+        r"""Returns the values stored in the field as an ndarray.
+
+        Parameters
+        ----------
+        flatten: bool
+            When it is True a flat ndarray is returned. Otherwise an ndarray with the field's
+            :obj:`shape` is returned.
+        dtype: str or dtype
+            Typecode or data-type to which the array is cast.
+
+        Returns
+        -------
+        ndarray
+            Field values
+
+        """
+        values = self.values
+        if not flatten:
+            values = self.values.reshape(self.shape)
+        if dtype is not None:
+            values = values.astype(dtype)
+        return values
+
+    def data(self, keys=("lat", "lon", "value"), flatten=False):
+        r"""Returns the values and/or the geographical coordinates for each grid point.
+
+        Parameters
+        ----------
+        keys: :obj:`str`, :obj:`list` or :obj:`tuple`
+            Specifies the type of data to be returned. Any combination of "lat", "lon" and "value"
+            is allowed here.
+        flatten: bool
+            When it is True a flat ndarray per key is returned. Otherwise an ndarray with the field's
+            :obj:`shape` is returned for each key.
+
+        Returns
+        -------
+        ndarray or tuple of ndarrays
+            When ``keys`` is a single value an ndarray is returned. Otherwise a tuple containing one ndarray
+            per key is returned (following the order in ``keys``).
+
+        See Also
+        --------
+        to_points
+        to_numpy
+        values
+
+        """
+        _keys = dict(
+            lat=self._metadata.latitudes,
+            lon=self._metadata.longitudes,
+            value=lambda: self.values,
+        )
+
+        if isinstance(keys, str):
+            keys = [keys]
+
+        for k in keys:
+            if k not in _keys:
+                raise ValueError(f"data: invalid argument: {k}")
+
+        r = [_keys[k]() for k in keys]
+        if not flatten:
+            shape = self.shape
+            r = [x.reshape(shape) for x in r]
+        return r[0] if len(r) == 1 else tuple(r)
+
+    def to_points(self, flatten=False):
+        r"""Returns the geographical coordinates in the data's original
+        Coordinate Reference System (CRS).
+
+        Parameters
+        ----------
+        flatten: bool
+            When it is True 1D ndarrays are returned. Otherwise ndarrays with the field's
+            :obj:`shape` are returned.
+
+        Returns
+        -------
+        dict
+            Dictionary with items "x" and "y", containing the ndarrays of the x and
+            y coordinates, respectively.
+
+        Raises
+        ------
+        ValueError
+            When the coordinates in the data's original CRS are not available.
+
+        See Also
+        --------
+        to_latlon
+
+        """
+        x = self._metadata.x()
+        y = self._metadata.y()
+        if x is not None and y is not None:
+            if not flatten:
+                shape = self.shape
+                x = x.reshape(shape)
+                y = y.reshape(shape)
+            return dict(x=x, y=y)
+        elif self.projection().CARTOPY_CRS == "PlateCarree":
+            lon, lat = self.data(("lon", "lat"), flatten=flatten)
+            return dict(x=lon, y=lat)
+        else:
+            raise ValueError(
+                "to_points(): geographical coordinates in original CRS are not available"
+            )
+
+    def to_latlon(self, flatten=False):
+        r"""Returns the latitudes/longitudes of all the gridpoints in the field.
+
+        Parameters
+        ----------
+        flatten: bool
+            When it is True 1D ndarrays are returned. Otherwise ndarrays with the field's
+            :obj:`shape` are returned.
+
+        Returns
+        -------
+        dict
+            Dictionary with items "lat" and "lon", containing the ndarrays of the latitudes and
+            longitudes, respectively.
+
+        See Also
+        --------
+        to_points
+
+        """
+        lon, lat = self.data(("lon", "lat"), flatten=flatten)
+        return dict(lat=lat, lon=lon)
+
+    @property
+    def shape(self):
+        r"""tuple: Gets the shape of the field. For structured grids the shape is a tuple
+        in the form of (Nj, Ni) where:
+
+        - ni: the number of gridpoints in i direction (longitude for a regular latitude-longitude grid)
+        - nj: the number of gridpoints in j direction (latitude for a regular latitude-longitude grid)
+
+        For other grid types the number of gridpoints is returned as ``(num,)``
+        """
+        return self._metadata.shape()
+
+    def projection(self):
+        r"""Returns information about the projection.
+
+        Returns
+        -------
+        :obj:`Projection`
+
+        Examples
+        --------
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/test.grib")
+        >>> ds.projection()
+        <Projected CRS: +proj=eqc +ellps=WGS84 +a=6378137.0 +lon_0=0.0 +to ...>
+        Name: unknown
+        Axis Info [cartesian]:
+        - E[east]: Easting (unknown)
+        - N[north]: Northing (unknown)
+        - h[up]: Ellipsoidal height (metre)
+        Area of Use:
+        - undefined
+        Coordinate Operation:
+        - name: unknown
+        - method: Equidistant Cylindrical
+        Datum: Unknown based on WGS 84 ellipsoid
+        - Ellipsoid: WGS 84
+        - Prime Meridian: Greenwich
+        >>> ds.projection().to_proj_string()
+        '+proj=eqc +ellps=WGS84 +a=6378137.0 +lon_0=0.0 +to_meter=111319.4907932736 +no_defs +type=crs'
+        """
+        return self._metadata.projection()
+
+    def bounding_box(self):
+        r"""Returns the bounding box of the field.
+
+        Returns
+        -------
+        :obj:`BoundingBox <data.utils.bbox.BoundingBox>`
+        """
+        return self._metadata.bounding_box()
+
+    def datetime(self):
+        r"""Returns the date and time of the field.
+
+        Returns
+        -------
+        dict of datatime.datetime
+            Dict with items "base_time" and "valid_time".
+
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "tests/data/t_time_series.grib")
+        >>> ds[4].datetime()
+        {'base_time': datetime.datetime(2020, 12, 21, 12, 0),
+        'valid_time': datetime.datetime(2020, 12, 21, 18, 0)}
+
+        """
+        return self._metadata.datetime()
+
+    def metadata(self, *keys, namespace=None, astype=None, **kwargs):
+        # when called without arguments returns the metadata object
+        if len(keys) == 0 and namespace is None and astype is None and len(kwargs) == 0:
+            return self._metadata
+
+        key, namespace, astype, key_arg_type = metadata_argument(
+            *keys, namespace=namespace, astype=astype
+        )
+
+        assert isinstance(key, list)
+        assert isinstance(namespace, (list, tuple))
+
+        if key:
+            assert isinstance(astype, (list, tuple))
+            if namespace and namespace[0] != "default":
+                key = [namespace[0] + "." + k for k in key]
+
+            r = [
+                self._metadata._get(k, astype=kt, **kwargs)
+                for k, kt in zip(key, astype)
+            ]
+
+            if key_arg_type == str:
+                return r[0]
+            elif key_arg_type == tuple:
+                return tuple(r)
+            else:
+                return r
+        elif namespace:
+            if all in namespace:
+                namespace = self._metadata.namespaces().keys()
+
+            r = {
+                ns: self._metadata.as_namespace(self._metadata.namespaces().get(ns, ns))
+                for ns in namespace
+            }
+            if len(r) == 1:
+                return r[namespace[0]]
+            else:
+                return r
+
+    def dump(self, namespace=None, **kwargs):
+        r"""Generates dump with all the metadata keys belonging to ``namespace``
+        offering a tabbed interface in a Jupyter notebook.
+
+        Parameters
+        ----------
+        namespace: :obj:`str`, :obj:`list` or :obj:`tuple`
+            The namespace to dump. Any :xref:`eccodes_namespace` can be used here.
+            :obj:`dump` also defines the "default" namespace, which contains all the GRIB keys
+            that ecCodes can access without specifying a namespace.
+            When ``namespace`` is None all the available namespaces will be used.
+        **kwargs: dict, optional
+            Other keyword arguments used for testing only
+
+        Returns
+        -------
+        NamespaceDump
+            Dict-like object with one item per namespace. In a Jupyter notebook represented
+            as a tabbed interface to browse the dump contents.
+
+        Examples
+        --------
+        :ref:`/examples/grib_metadata.ipynb`
+
+        """
+        return self._metadata.dump(namespace=namespace, **kwargs)
+
+    def __getitem__(self, key):
+        """Returns the value of the metadata ``key``."""
+        return self._metadata.get(key)
+
+    @abstractmethod
+    def message(self):
+        r"""Returns a buffer containing the encoded message for message based formats (e.g. GRIB).
+
+        Returns
+        -------
+        bytes
+        """
+        self._not_implemented()
+
+    @abstractmethod
+    def _attributes(self, names):
+        result = {}
+        for name in names:
+            result[name] = self._metadata.get(name, None)
+        return result
+
+
+class FieldList(Index):
+    _indices = {}
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    @staticmethod
+    def from_numpy(array, metadata):
+        from earthkit.data.sources.numpy_list import NumpyFieldList
+
+        return NumpyFieldList(array, metadata)
+
+    @cached_method
+    def _default_index_keys(self):
+        if len(self) > 0:
+            return self[0].metadata().index_keys()
+        else:
+            return []
+
+    def _find_index_values(self, key):
+        values = set()
+        for f in self:
+            v = f.metadata(key, default=None)
+            if v is not None:
+                values.add(v)
+        return sorted(list(values))
+
+    def _find_all_index_dict(self):
+        indices = defaultdict(set)
+        for f in self:
+            for k in self._default_index_keys():
+                v = f.metadata(k, default=None)
+                if v is None:
+                    continue
+                indices[k].add(v)
+
+        return {k: sorted(list(v)) for k, v in indices.items()}
+
+    def indices(self, squeeze=False):
+        r"""Returns the unique, sorted values for a set of metadata keys (see below)
+        from all the fields. Individual keys can be also queried by :obj:`index`.
+
+        Parameters
+        ----------
+        squeeze : False
+            When True only returns the metadata keys that have more than one values.
+
+        Returns
+        -------
+        dict
+            Unique, sorted metadata values from all the
+            :obj:`GribField <data.readers.grib.codes.GribField>`\ s.
+
+        See Also
+        --------
+        index
+
+        Examples
+        --------
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/tuv_pl.grib")
+        >>> ds.indices()
+        {'class': ['od'], 'stream': ['oper'], 'levtype': ['pl'], 'type': ['an'],
+        'expver': ['0001'], 'date': [20180801], 'time': [1200], 'domain': ['g'],
+        'number': [0], 'levelist': [300, 400, 500, 700, 850, 1000],
+        'param': ['t', 'u', 'v']}
+        >>> ds.indices(squeeze=True)
+        {'levelist': [300, 400, 500, 700, 850, 1000], 'param': ['t', 'u', 'v']}
+        >>> ds.index("level")
+        [300, 400, 500, 700, 850, 1000]
+
+        By default :obj:`indices` uses the keys from the
+        "mars" :xref:`eccodes_namespace`. Keys with no valid values are not included. Keys
+        that :obj:`index` is called with are automatically added to the original set of keys
+        used in :obj:`indices`.
+
+        """
+        if not self._indices:
+            self._indices = self._find_all_index_dict()
+        if squeeze:
+            return {k: v for k, v in self._indices.items() if len(v) > 1}
+        else:
+            return self._indices
+
+    def index(self, key):
+        r"""Returns the unique, sorted values of the specified metadata ``key`` from all the fields.
+        ``key`` will be automatically added to the keys returned by :obj:`indices`.
+
+        Parameters
+        ----------
+        key : str
+            Metadata key.
+
+        Returns
+        -------
+        list
+            Unique, sorted values of ``key`` from all the
+            :obj:`GribField <data.readers.grib.codes.GribField>`\ s.
+
+        See Also
+        --------
+        index
+
+        Examples
+        --------
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/tuv_pl.grib")
+        >>> ds.index("level")
+        [300, 400, 500, 700, 850, 1000]
+
+        """
+        if key in self.indices():
+            return self.indices()[key]
+
+        self._indices[key] = self._find_index_values(key)
+        return self._indices[key]
+
+    def to_numpy(self, **kwargs):
+        r"""Returns the field values as an ndarray. It is formed as the array of the
+        :obj:`data.core.fieldlist.Field.to_numpy` values per field.
+
+        Parameters
+        ----------
+        **kwargs: dict, optional
+            Keyword arguments passed to :obj:`data.core.fieldlist.Field.to_numpy`
+
+        Returns
+        -------
+        ndarray
+            Array containing the field values.
+
+        See Also
+        --------
+        values
+        """
+        import numpy as np
+
+        return np.array([f.to_numpy(**kwargs) for f in self])
+
+    @property
+    def values(self):
+        r"""ndarray: Gets the field values as a 2D ndarray. It is formed as the array of
+        :obj:`GribField.values <data.readers.grib.codes.GribField.values>` per field.
+
+        See Also
+        --------
+        to_numpy
+
+
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/test.grib")
+        >>> for f in ds:
+        ...     print(f.values.shape)
+        ...
+        (209,)
+        (209,)
+        >>> v = ds.values
+        >>> v.shape
+        (2, 209)
+        >>> v[0][:3]
+        array([262.78027344, 267.44726562, 268.61230469])
+
+        """
+        import numpy as np
+
+        return np.array([f.values for f in self])
+
+    def metadata(self, *args, **kwargs):
+        r"""Returns the metadata values for each field.
+
+        Parameters
+        ----------
+        *args: tuple
+            Positional arguments defining the metadata keys. Passed to
+            :obj:`GribField.metadata() <data.readers.grib.codes.GribField.metadata>`
+        **kwargs: dict, optional
+            Keyword arguments passed to
+            :obj:`GribField.metadata() <data.readers.grib.codes.GribField.metadata>`
+
+        Returns
+        -------
+        list
+            List with one item per :obj:`GribField <data.readers.grib.codes.GribField>`
+
+        Examples
+        --------
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/test.grib")
+        >>> ds.metadata("param")
+        ['2t', 'msl']
+        >>> ds.metadata("param", "units")
+        [('2t', 'K'), ('msl', 'Pa')]
+        >>> ds.metadata(["param", "units"])
+        [['2t', 'K'], ['msl', 'Pa']]
+
+        """
+        result = []
+        for s in self:
+            result.append(s.metadata(*args, **kwargs))
+        return result
+
+    def ls(self, n=None, keys=None, extra_keys=None, namespace=None):
+        from earthkit.data.utils.summary import ls
+
+        def _proc(keys, n):
+            num = len(self)
+            pos = slice(0, num)
+            if n is not None:
+                pos = slice(0, min(num, n)) if n > 0 else slice(num - min(num, -n), num)
+            pos_range = range(pos.start, pos.stop)
+
+            if "namespace" in keys:
+                ns = keys.pop("namespace", None)
+                for i in pos_range:
+                    f = self[i]
+                    v = f.metadata(namespace=ns)
+                    if len(keys) > 0:
+                        v.update(f._attributes(keys))
+                    yield (v)
+            else:
+                for i in pos_range:
+                    yield (self[i]._attributes(keys))
+
+        _keys = (
+            self._default_ls_keys() if namespace is None else dict(namespace=namespace)
+        )
+        return ls(_proc, _keys, n=n, keys=keys, extra_keys=extra_keys)
+
+    @cached_method
+    def _default_ls_keys(self):
+        if len(self) > 0:
+            return self[0].metadata().ls_keys()
+        else:
+            return []
+
+    def head(self, n=5, **kwargs):
+        r"""Generates a list like summary of the first ``n``
+        :obj:`GribField <data.readers.grib.codes.GribField>`\ s using a set of metadata keys.
+        Same as calling :obj:`ls` with ``n``.
+
+        Parameters
+        ----------
+        n: int, None
+            The number of messages (``n`` > 0) to be printed from the front.
+        **kwargs: dict, optional
+            Other keyword arguments passed to :obj:`ls`.
+
+        Returns
+        -------
+        Pandas DataFrame
+            See  :obj:`ls`.
+
+
+        The following calls are equivalent:
+
+            .. code-block:: python
+
+                ds.head()
+                ds.head(5)
+                ds.head(n=5)
+                ds.ls(5)
+                ds.ls(n=5)
+
+        """
+        if n <= 0:
+            raise ValueError("head: n must be > 0")
+        return self.ls(n=n, **kwargs)
+
+    def tail(self, n=5, **kwargs):
+        r"""Generates a list like summary of the last ``n``
+        :obj:`GribField <data.readers.grib.codes.GribField>`\ s using a set of metadata keys.
+        Same as calling :obj:`ls` with ``-n``.
+
+        Parameters
+        ----------
+        n: int, None
+            The number of messages (``n`` > 0)  to be printed from the back.
+        **kwargs: dict, optional
+            Other keyword arguments passed to :obj:`ls`.
+
+        Returns
+        -------
+        Pandas DataFrame
+            See  :obj:`ls`.
+
+
+        The following calls are equivalent:
+
+            .. code-block:: python
+
+                ds.tail()
+                ds.tail(5)
+                ds.tail(n=5)
+                ds.ls(-5)
+                ds.ls(n=-5)
+
+        """
+        if n <= 0:
+            raise ValueError("n must be > 0")
+        return self.ls(n=-n, **kwargs)
+
+    def describe(self, *args, **kwargs):
+        r"""Generates a summary of the fieldlist."""
+        from earthkit.data.utils.summary import format_describe
+
+        def _proc():
+            for f in self:
+                yield (f._attributes(self._describe_keys()))
+
+        return format_describe(_proc(), *args, **kwargs)
+
+    @cached_method
+    def _describe_keys(self):
+        if len(self) > 0:
+            return self[0].metadata().describe_keys()
+        else:
+            return []
+
+    def datetime(self):
+        r"""Returns the unique, sorted list of dates and times in the fieldlist.
+
+        Returns
+        -------
+        dict of datatime.datetime
+            Dict with items "base_time" and "valid_time".
+
+
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "tests/data/t_time_series.grib")
+        >>> ds.datetime()
+        {'base_time': [datetime.datetime(2020, 12, 21, 12, 0)],
+        'valid_time': [
+            datetime.datetime(2020, 12, 21, 12, 0),
+            datetime.datetime(2020, 12, 21, 15, 0),
+            datetime.datetime(2020, 12, 21, 18, 0),
+            datetime.datetime(2020, 12, 21, 21, 0),
+            datetime.datetime(2020, 12, 23, 12, 0)]}
+
+        """
+        base = set()
+        valid = set()
+        for s in self:
+            d = s.datetime()
+            base.add(d["base_time"])
+            valid.add(d["valid_time"])
+        return {"base_time": sorted(base), "valid_time": sorted(valid)}
+
+    def to_points(self, **kwargs):
+        r"""Returns the geographical coordinates shared by all the fields in
+        the data's original Coordinate Reference System (CRS).
+
+        Parameters
+        ----------
+        **kwargs: dict, optional
+            Keyword arguments passed to
+            :obj:`GribField.to_points() <data.readers.grib.codes.GribField.to_points>`
+
+        Returns
+        -------
+        dict
+            Dictionary with items "x" and "y", containing the ndarrays of the x and
+            y coordinates, respectively.
+
+        Raises
+        ------
+        ValueError
+            When not all the fields have the same grid geometry.
+        """
+        if self._is_shared_grid():
+            return self[0].to_points(**kwargs)
+        elif len(self) == 0:
+            return dict(x=None, y=None)
+        else:
+            raise ValueError("Fields do not have the same grid geometry")
+
+    def to_latlon(self, **kwargs):
+        r"""Returns the latitudes/longitudes shared by all the fields.
+
+        Parameters
+        ----------
+        **kwargs: dict, optional
+            Keyword arguments passed to
+            :obj:`GribField.to_latlon() <data.readers.grib.codes.GribField.to_latlon>`
+
+        Returns
+        -------
+        dict
+            Dictionary with items "lat" and "lon", containing the ndarrays of the latitudes and
+            longitudes, respectively.
+
+        Raises
+        ------
+        ValueError
+            When not all the fields have the same grid geometry
+
+        Examples
+        --------
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/test.grib")
+        >>> for f in ds:
+        ...     print(f.shape)
+        ...
+        (11, 19)
+        (11, 19)
+        >>> r = ds.to_latlon()
+        >>> for k, v in r.items():
+        ...     print(f"{k}: shape={v.shape}")
+        ...
+        lat: shape=(11, 19)
+        lon: shape=(11, 19)
+        >>> r["lon"][:2]
+        array([[-27., -23., -19., -15., -11.,  -7.,  -3.,   1.,   5.,   9.,  13.,
+         17.,  21.,  25.,  29.,  33.,  37.,  41.,  45.],
+        [-27., -23., -19., -15., -11.,  -7.,  -3.,   1.,   5.,   9.,  13.,
+         17.,  21.,  25.,  29.,  33.,  37.,  41.,  45.]])
+
+        """
+        if self._is_shared_grid():
+            return self[0].to_latlon(**kwargs)
+        elif len(self) == 0:
+            return dict(lat=None, lon=None)
+        else:
+            raise ValueError("Fields do not have the same grid geometry")
+
+    def projection(self):
+        r"""Returns the projection information shared by all the fields.
+
+        Returns
+        -------
+        :obj:`Projection`
+
+        Raises
+        ------
+        ValueError
+            When not all the fields have the same grid geometry
+
+        Examples
+        --------
+        >>> import earthkit.data
+        >>> ds = earthkit.data.from_source("file", "docs/examples/test.grib")
+        >>> ds.projection()
+        <Projected CRS: +proj=eqc +ellps=WGS84 +a=6378137.0 +lon_0=0.0 +to ...>
+        Name: unknown
+        Axis Info [cartesian]:
+        - E[east]: Easting (unknown)
+        - N[north]: Northing (unknown)
+        - h[up]: Ellipsoidal height (metre)
+        Area of Use:
+        - undefined
+        Coordinate Operation:
+        - name: unknown
+        - method: Equidistant Cylindrical
+        Datum: Unknown based on WGS 84 ellipsoid
+        - Ellipsoid: WGS 84
+        - Prime Meridian: Greenwich
+        >>> ds.projection().to_proj_string()
+        '+proj=eqc +ellps=WGS84 +a=6378137.0 +lon_0=0.0 +to_meter=111319.4907932736 +no_defs +type=crs'
+        """
+        if self._is_shared_grid():
+            return self[0].projection()
+        elif len(self) == 0:
+            return None
+        else:
+            raise ValueError("Fields do not have the same grid geometry")
+
+    def bounding_box(self):
+        r"""Returns the bounding box for each field.
+
+        Returns
+        -------
+        list
+            List with one :obj:`BoundingBox <data.utils.bbox.BoundingBox>` per
+            :obj:`GribField <data.readers.grib.codes.GribField>`
+        """
+        return [s.bounding_box() for s in self]
+
+    @cached_method
+    def _is_shared_grid(self):
+        if len(self) > 0:
+            grid = self[0].metadata()._unique_grid_id()
+            if grid is not None:
+                return all(f.metadata()._unique_grid_id() == grid for f in self)
+        return False
+
+    def save(self, filename):
+        r"""Writes all the fields into a file. The target file will be overwritten if
+        already exists.
+
+        Parameters
+        ----------
+        filename: str
+            The target file path.
+        """
+        with open(filename, "wb") as f:
+            self.write(f)
+
+    def write(self, f):
+        r"""Writes all the fields to a file object.
+
+        Parameters
+        ----------
+        f: file object
+            The target file object.
+        """
+        for s in self:
+            s.write(f)
