+++ earthkit/data/readers/grib/index/__init__.py
@@ -12,6 +12,7 @@ import math
 import os
 from abc import abstractmethod
 
+from earthkit.data.core.fieldlist import FieldList
 from earthkit.data.core.index import Index, MaskIndex, MultiIndex
 from earthkit.data.decorators import alias_argument
 from earthkit.data.indexing.database import (
@@ -21,14 +22,15 @@ from earthkit.data.indexing.database import (
     STATISTICS_KEY_NAMES,
 )
 from earthkit.data.readers.grib.codes import GribField
-from earthkit.data.readers.grib.fieldlist import FieldListMixin
+from earthkit.data.readers.grib.pandas import PandasMixIn
+from earthkit.data.readers.grib.xarray import XarrayMixIn
 from earthkit.data.utils import progress_bar
 from earthkit.data.utils.availability import Availability
 
 LOG = logging.getLogger(__name__)
 
 
-class FieldList(FieldListMixin, Index):
+class GribFieldList(PandasMixIn, XarrayMixIn, FieldList):
     r"""Represents a list of :obj:`GribField <data.readers.grib.codes.GribField>`\ s.
 
     We can **iterate** through the fields as follows:
@@ -111,7 +113,7 @@ class FieldList(FieldListMixin, Index):
 
     @classmethod
     def new_mask_index(self, *args, **kwargs):
-        return MaskFieldList(*args, **kwargs)
+        return GribMaskFieldList(*args, **kwargs)
 
     @property
     def availability_path(self):
@@ -119,8 +121,8 @@ class FieldList(FieldListMixin, Index):
 
     @classmethod
     def merge(cls, sources):
-        assert all(isinstance(_, FieldList) for _ in sources)
-        return MultiFieldList(sources)
+        assert all(isinstance(_, GribFieldList) for _ in sources)
+        return GribMultiFieldList(sources)
 
     def _custom_availability(self, ignore_keys=None, filter_keys=lambda k: True):
         def dicts():
@@ -178,17 +180,17 @@ class FieldList(FieldListMixin, Index):
         return kwargs
 
 
-class MaskFieldList(FieldList, MaskIndex):
+class GribMaskFieldList(GribFieldList, MaskIndex):
     def __init__(self, *args, **kwargs):
         MaskIndex.__init__(self, *args, **kwargs)
 
 
-class MultiFieldList(FieldList, MultiIndex):
+class GribMultiFieldList(GribFieldList, MultiIndex):
     def __init__(self, *args, **kwargs):
         MultiIndex.__init__(self, *args, **kwargs)
 
 
-class FieldListInFiles(FieldList):
+class GribFieldListInFiles(GribFieldList):
     # Remote FieldLists (with urls) are also here,
     # as the actual fieldlist is accessed on a file in cache.
     # This class changes the interface (_getitem__ and __len__)
