+++ earthkit/data/readers/grib/index/db.py
@@ -13,13 +13,13 @@ import os
 from urllib.parse import urljoin
 
 from earthkit.data.core.caching import cache_file
-from earthkit.data.readers.grib.index import FieldListInFiles
+from earthkit.data.readers.grib.index import GribFieldListInFiles
 from earthkit.data.utils import progress_bar
 
 LOG = logging.getLogger(__name__)
 
 
-class FieldListInFilesWithDBIndex(FieldListInFiles):
+class FieldListInFilesWithDBIndex(GribFieldListInFiles):
     def __init__(self, db, **kwargs):
         """Should not be instantiated directly.
         The public API are the constructors "_from*()" class methods.
